refactor: replace `semaphore-async-await` dependency with smaller implementation (#2222)
